Refactor constructor view + fixes
