+++ src/views/Constructor/components/education.vue
